requirements.txt
django
pillow
djangorestframework
drf-yasg
djangorestframework-simplejwt
django-bootstrap-v5